tests: parametrize reset/backup tests
